+++ src/nreplController.ts
@@ -1,7 +1,8 @@
 import 'process';
+import * as os from 'os';
 import * as vscode from 'vscode';
 import { spawn } from 'cross-spawn';
-import { ChildProcess } from 'child_process';
+import { ChildProcess, exec } from 'child_process';
 
 import { CljConnectionInformation } from './cljConnection';
 
@@ -10,6 +11,11 @@ const LEIN_ARGS: string[] = [
     ':dependencies',
     'conj',
     '[org.clojure/tools.nrepl "0.2.12" :exclusions [org.clojure/clojure]]',
+    '--',
+    'update-in',
+    ':dependencies',
+    'conj',
+    '[cljfmt "0.5.7"]',
     '--',
     'update-in',
     ':plugins',
@@ -34,7 +40,10 @@ const start = (): Promise<CljConnectionInformation> => {
     if (isStarted())
         return Promise.reject({ nreplError: 'nREPL already started.' });
 
-    nreplProcess = spawn('lein', LEIN_ARGS, { cwd: vscode.workspace.rootPath, detached: true });
+    nreplProcess = spawn('lein', LEIN_ARGS, {
+        cwd: vscode.workspace.rootPath,
+        detached: !(os.platform() === 'win32')
+    });
 
     return new Promise((resolve, reject) => {
         nreplProcess.stdout.addListener('data', data => {
@@ -68,7 +77,12 @@ const stop = () => {
     if (nreplProcess) {
         // Workaround http://azimi.me/2014/12/31/kill-child_process-node-js.html
         nreplProcess.removeAllListeners();
-        process.kill(-nreplProcess.pid);
+        if(os.platform() === 'win32'){
+            exec('taskkill /pid ' + nreplProcess.pid + ' /T /F')
+        }
+        else {
+            process.kill(-nreplProcess.pid);
+        }
         nreplProcess = null;
     }
 };